Zmiana belki w trakcie konkursu + (patrz opis)
- Możliwość włączenia/wyłączenia w ustawieniach opcji pokazywania numerów startowych zawodników przy pokazywaniu wyników lub zapisywanie ich do pliku. (domyślnie wyłączone)
- Aby zmienić belkę w trakcie konkursu, należy wcisnąć 'b' i podać belkę.

+++ src/Hill.h
@@ -16,7 +16,7 @@ private:
     double pointsForKPoint, pointsForMeters;
 
     double startDistance;
-    int landingDifficulty;
+    double landingDifficulty;
 
     double takeoffTechniqueMeters, flightTechniqueMeters;
     int optimalTakeoffPower;
